【rpm生成】修改spec文件，重新生成rpm包

+++ rpmbuild-x86/rpmbuild/SPECS/hellorpm.spec
@@ -8,7 +8,7 @@ Summary:        This is a test of rpm build
 
 Group:          Development/Libraries
 License:        GPL 
-URL:            https://github.com:swiftchao/rpm/rpmbuild-aarch/rpmbuild
+URL:            https://github.com:swiftchao/rpm
 Source0:        %{name}-%{version}-%{release}.%{_arch}.tar.gz 
 #将patch文件放在rpmbuild/SOURCES目录下
 #在Sourcexx结束和Buildrequires之前加入patch，一般未使用补丁号按递增命名
@@ -19,49 +19,49 @@ Patch0:0001-test-patch.patch
 %{summary}
 
 %prep
-#cur=/root/rpmbuild/BUILD
-cd $RPM_SOURCE_DIR
-tar -xvf $RPM_SOURCE_DIR/%{name}-%{version}-%{release}.%{_arch}.tar.gz
-
 %setup -n %{name}-%{version}-%{release}.%{_arch}
 #在setup后build之前加入使用patch
 %patch0 -p4
 
 %build
-cd $RPM_SOURCE_DIR/%{name}-%{version}-%{release}.%{_arch}
 %configure
 make clean
 make %{?_smp_mflags}
 
 %install
 rm -rf $RPM_BUILD_ROOT
-cd $RPM_SOURCE_DIR/%{name}-%{version}-%{release}.%{_arch}
-make install
-cp -r $RPM_SOURCE_DIR/%{name}-%{version}-%{release}.%{_arch} %{_topdir}/BUILDROOT/
+echo $RPM_BUILD_ROOT
+mkdir -p $RPM_BUILD_ROOT/usr/bin
+mkdir -p $RPM_BUILD_ROOT/usr/local/share
+cp -f $RPM_BUILD_DIR/%{name}-%{version}-%{release}.%{_arch}/test-main $RPM_BUILD_ROOT/usr/bin/
+cp -r $RPM_BUILD_DIR/%{name}-%{version}-%{release}.%{_arch} $RPM_BUILD_ROOT/usr/local/share/
+
 
 %post
+chmod a+x /usr/bin/test-main
 mkdir -p ${HOME}/bin
 echo mkdir -p ${HOME}/bin >>/tmp/chaofeirpm.log
-cp test-main ${HOME}/bin/test-main
+cp /usr/bin/test-main ${HOME}/bin/test-main
 chmod a+x ${HOME}/bin/test-main
-cp test-main /usr/bin/test-main
-chmod a+x /usr/bin/test-main
 echo ${HOME}/bin/test-main
 ${HOME}/bin/test-main
 echo ${HOME}/bin/test-main >>/tmp/chaofeirpm.log
 
 %postun
-rm -f ${HOME}/bin/test-main
-rm -f /usr/bin/test-main
-rm -f /tmp/chaofeirpm.log
+if [ $1 = 0 ]; then
+  rm -f ${HOME}/bin/test-main
+  rm -f /usr/bin/test-main
+  rm -f /tmp/chaofeirpm.log
+fi
 
 %clean
 rm -rf $RPM_BUILD_ROOT
+make clean
 
 %files
 %defattr(-,root,root,-)
-/*
-#/test-main
+/usr/local/share/%{name}-%{version}-%{release}.%{_arch}
+/usr/bin/test-main
 
 %changelog * Wed Mar 1  2019 Chao Fei <chaofeibest@163.com> -0.0.1-1
